Start 3 servers for 3 GraphQL schemas and add configurations for deployment
