+++ charts/galoy/templates/api-ingress.yaml
@@ -1,10 +1,10 @@
-{{ if .Values.adminIngress }}
+{{ if .Values.apiIngress }}
 apiVersion: networking.k8s.io/v1beta1
 kind: Ingress
 metadata:
-  name: {{ .Values.adminIngress.name }}
+  name: {{ .Values.apiIngress.name }}
   labels:
-    app: {{ .Values.adminIngress.name }}
+    app: {{ .Values.apiIngress.name }}
     app.kubernetes.io/managed-by: Helm
   annotations:
     kubernetes.io/ingress.class: nginx
@@ -13,14 +13,14 @@ metadata:
 spec:
   tls:
   - hosts:
-    - {{.Values.adminIngress.host}}
-    secretName: {{ .Values.adminIngress.name }}-tls
+    - {{.Values.apiIngress.host}}
+    secretName: {{.Values.apiIngress.name}}-tls
   rules:
-  - host: {{.Values.adminIngress.host}}
+  - host: {{.Values.apiIngress.host}}
     http:
       paths:
       - path: /
         backend:
-          serviceName: graphql-admin
-          servicePort: 4001
+          serviceName: api
+          servicePort: 4002
 {{ end }}
