+++ package.json
@@ -7,7 +7,7 @@
     "prettier-check": "prettier --check .",
     "prettier-write": "prettier --write .",
     "build": "tsc -p tsconfig-build.json --skipLibCheck && cp -R src/services/protos lib/services/ && cp src/graphql/old-schema.graphql lib/graphql/ && tscpaths -p tsconfig.json -s ./src -o ./lib",
-    "start": "yarn build && source ./scripts/export-local.sh && node lib/servers/graphql-main-server.js | pino-pretty -c -l",
+    "start": "yarn build && source ./scripts/export-local.sh && node lib/servers/graphql-old-server.js | pino-pretty -c -l",
     "trigger": "yarn build && source ./scripts/export-local.sh && node lib/servers/trigger.js | pino-pretty -c -l",
     "watch": "nodemon -V -e ts,graphql -w ./src -x yarn run start",
     "watch-trigger": "nodemon -V -e ts,graphql -w ./src -x yarn trigger",
@@ -25,7 +25,6 @@
     "postman": "yarn node test/postman/newman.ts",
     "bitcoin-cli": "source ./scripts/export-local.sh && kubectl exec bitcoind-0 -- bitcoin-cli -rpcuser=rpcuser -rpcpassword=$BITCOINDRPCPASS -regtest",
     "update-postman": "curl --location --request GET \"https://api.getpostman.com/collections/4e3fee88-7267-44e1-a316-b81e90b0bd2b\" --header \"X-Api-Key: $POSTMANAPIKEY\" | jq . > test/postman/graphql.postman-collection.json",
-    "autocannon-rn-no-auth": "yarn autocannon https://graphql.testnet.galoy.io:443/graphql -m POST -H Content-Type=application/json -b '{\"query\":\"query gql_query_anonymous {\n  prices {\n    __typename\n    id\n    o\n  }\n  earnList {\n    __typename\n    id\n    value\n  }\n  maps {\n    __typename\n    id\n    title\n    coordinate {\n        __typename\n        latitude\n        longitude\n    }\n  }\n}\",\"variables\":{}}' -c 20 -d 10",
     "test": "export JEST_JUNIT_OUTPUT_DIR=\"./artifacts\" && jest --ci --reporters=default --reporters=jest-junit",
     "test-local": "yarn use-minikube && source ./scripts/export-local.sh && yarn jest --runInBand --verbose $TEST | pino-pretty -c -l",
     "test-local-restart": "./scripts/delete-pods.sh && source ./scripts/export-local.sh && yarn run test-local",
